Refactored routing
