--- server/backend/src/route/home.route.ts
@@ -1,10 +1,0 @@
-import express from 'express';
-import homeCtrl from '../api/home';
-const router = express.Router();
-
-router.route('/')
-    .get(homeCtrl.apiGetHome)
-    .post(homeCtrl.apiPostHome)
-    .put(homeCtrl.apiPutHome)
-
-export default router;
