--- server/backend/src/route/user.route.ts
@@ -1,10 +1,0 @@
-import express from 'express';
-import userCtrl from '../api/user';
-const router = express.Router();
-
-router.route('/')
-    .get(userCtrl.apiGetUser)
-    .post(userCtrl.apiPostUser)
-    .put(userCtrl.apiPutUser)
-
-export default router;
